[ADD] #6 - 코드리뷰 반영
